+++ src/Controller/AssetsController.php
@@ -35,6 +35,12 @@ class AssetsController extends Controller
             throw new AssetNotFoundException(__d('assets', 'File `{0}` doesn\'t exist', $file));
         }
 
-        return $this->response->withFile($file)->withType(pathinfo($file, PATHINFO_EXTENSION));
+        $response = $this->response->withModified(filemtime($file));
+
+        if ($response->checkNotModified($this->request)) {
+            return $response;
+        }
+
+        return $response->withFile($file)->withType(pathinfo($file, PATHINFO_EXTENSION));
     }
 }